+++ packages/common-all/package.json
@@ -43,7 +43,6 @@
     "fuse.js": "^6.4.6",
     "github-slugger": "^1.3.0",
     "gray-matter": "^4.0.2",
-    "handlebars": "^4.7.7",
     "http-status-codes": "^2.1.4",
     "isomorphic-fetch": "^2.2.1",
     "lodash": "^4.17.15",
