+++ packages/common-server/src/index.ts
@@ -10,3 +10,4 @@ export * from "./errorReporting";
 export * from "./types";
 export * from "./server";
 export * from "./yaml";
+export * from "./template";
